cleaning up old tests

+++ src/main/Security/EncryptionTools.java
@@ -72,20 +72,22 @@ public class EncryptionTools {
       System.err.println("Error creating primitive: %s " + ex);
       System.exit(1);
     }
-
-    KeysetHandle handle = KeysetHandle.generateNew(KeyTemplates.get("AES256_GCM"));
-    String keysetFileName = "key.json";
-    handle.write(JsonKeysetWriter.withPath(keysetFileName), kekAead);
-
-    ObjectMapper objectMapper = new ObjectMapper();
     File file = new File("key.json");
-    Map<String, Object> map =
-        objectMapper.readValue(file, new TypeReference<Map<String, Object>>() {});
-    JSONObject keyJson = new JSONObject(map);
+    if (!file.exists()) {
+      KeysetHandle handle = KeysetHandle.generateNew(KeyTemplates.get("AES256_GCM"));
+      String keysetFileName = "key.json";
+      handle.write(JsonKeysetWriter.withPath(keysetFileName), kekAead);
 
-    MongoCollection<Document> keyHandles = this.db.getCollection("keys", Document.class);
-    keyHandles.deleteMany(new Document());
-    //    Document keyDocument = new Document("key.json");
-    keyHandles.insertOne(Document.parse(keyJson.toString()));
+      ObjectMapper objectMapper = new ObjectMapper();
+      Map<String, Object> map =
+          objectMapper.readValue(file, new TypeReference<Map<String, Object>>() {
+          });
+      JSONObject keyJson = new JSONObject(map);
+
+      MongoCollection<Document> keyHandles = this.db.getCollection("keys", Document.class);
+      keyHandles.deleteMany(new Document());
+      //    Document keyDocument = new Document("key.json");
+      keyHandles.insertOne(Document.parse(keyJson.toString()));
+    }
   }
 }
